exit install.sh on error

+++ install.sh
@@ -6,6 +6,10 @@ source setup.sh
 #MadGraph
 wget https://launchpad.net/mg5amcnlo/3.0/3.1.x/+download/MG5_aMC_v3.1.1.tar.gz
 tar -xzvf MG5_aMC_v3.1.1.tar.gz
+if [[ $? -ne 0 ]]; then
+    echo "ERROR getting madgraph"
+    exit
+fi
 rm MG5_aMC_v3.1.1.tar.gz
 
 #Pythia
@@ -16,6 +20,10 @@ cd pythia8306
 ./configure --prefix=$baseDir
 make -j
 make install
+if [[ $? -ne 0 ]]; then
+    echo "ERROR compiling pythia"
+    exit
+fi
 cd ..
 
 #Delphes
@@ -24,5 +32,9 @@ export PYTHIA8=$baseDir
 git clone https://github.com/delphes/delphes.git
 cd delphes
 make HAS_PYTHIA8=true -j -I${baseDir}/include/Pythia8
+if [[ $? -ne 0 ]]; then
+    echo "ERROR compiling delphes"
+    exit
+fi
 
 #echo "export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:${baseDir}/lib" >> ~/.bash_profile
